Repair defective safe-allocation transformations
The original transformation (36cca4a2c75) left many of the touched
files uncompilable or inconsistent.
The following is repaired:

- safe_* calling malloc/calloc/abort without any declaration in scope
  (Juliet .i files and others; ~19k files): the safe_* method now sits
  after the declarations it needs. Missing extern declarations of the
  allocators and abort are added alongside it
- safe_* using size_t in files that never declare it (product-lines,
  array-fpi, ...): size_t is now declared after the malloc declaration
- renamed allocator prototypes (intel-tdx-module, array-fpi, ...): local
  declarations like 'extern void* malloc(size_t);' are kept unchanged so
  the safe_* body remains well-typed
- safe_* definitions accidentally nested inside functions (ldv model
  fragments; a GNU extension that clang and verifier frontends reject):
  fixed by moving the definitions.
- duplicate safe_* definitions in translation units composed via
  '#include "...c"' (memsafety-cve incl. todo/, ldv model families):
  exactly one definition per safe_* remains, placed before all uses
- misplaced safe_*: sqlite3.c (safe_* definition was inside an Apple-only #ifdef),
  ntdrivers SLAM models (safe_* now typed to match the file's own
  'char *malloc(int)' model and placed after that)

+++ c/verifythis/duplets.c
@@ -19,7 +19,7 @@ void __VERIFIER_assert(int cond) {
 
 extern int __VERIFIER_nondet_int(void);
 extern void abort(void);
-void *safe_calloc(size_t num, size_t size) {
+void *safe_calloc(unsigned int num, unsigned int size) {
   void *p = calloc(num, size);
   if (p == 0) {
     abort();
